@@ -1,11 +1,11 @@
 [project]
 name = "rispy"
-license = {file = "LICENSE"}
+license = { file = "LICENSE" }
 authors = [
-    {name = "Maik Derstappen", email = "md@derico.de"}
+    { name = "Maik Derstappen", email = "md@derico.de" }
 ]
 maintainers = [
-    {name = "Andy Shapiro", email = "shapiromatron@gmail.com"}
+    { name = "Andy Shapiro", email = "shapiromatron@gmail.com" }
 ]
 readme = "README.md"
 dynamic = ["version", "description"]
@@ -14,13 +14,13 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 
 [project.urls]
 "Source" = "https://github.com/mrtango/rispy"
@@ -33,7 +33,7 @@ dev = [
     "pytest ~=7.4.4",
     "pytest-benchmark",
     "flit ~= 3.9.0",
-    "ruff ~= 0.1.13",
+    "ruff ~= 0.11.6",
     "coverage ~= 7.4.0",
 ]
 
@@ -51,10 +51,12 @@ exclude = [".github", "Makefile", "tests"]
 
 [tool.ruff]
 line-length = 100
-target-version = "py38"
+target-version = "py39"
+
+[tool.ruff.lint]
 select = ["F", "E", "W", "I", "UP", "S", "B", "T20", "RUF"]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "test_*.py" = ["S101"]
 
 [tool.pytest.ini_options]
